updated based on requirement

+++ src/main/java/com/jpmc/theater/service/TheaterReservationService.java
@@ -6,6 +6,8 @@ import com.jpmc.theater.model.Customer;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.math.BigDecimal;
+
 @Service
 public class TheaterReservationService {
 
@@ -21,7 +23,7 @@ public class TheaterReservationService {
         return new Reservation(customer, showing, howManyTickets);
     }
 
-    public double getTotalFeeByReservation(Reservation reservation) {
+    public BigDecimal getTotalFeeByReservation(Reservation reservation) {
       return movieFeeCalculationService.calculateFee(reservation);
     }
