+++ src/test/java/com/jpmc/theater/service/MovieFeeCalculationServiceTest.java
@@ -23,7 +23,7 @@ public class MovieFeeCalculationServiceTest {
     public void specialMovieWith20PercentDiscount() {
         Movie spiderMan = new Movie("Spider-Man: No Way Home", Duration.ofMinutes(90),12.5, 1);
         Showing showing = new Showing(spiderMan, 5, LocalDateTime.of(LocalDate.now(), LocalTime.of(9, 0)));
-        assertEquals(10, movieFeeCalculationService.calculateTicketPrice(showing));
+        assertEquals(10.00, movieFeeCalculationService.calculateTicketPrice(showing).doubleValue());
     }
 
 
@@ -31,11 +31,11 @@ public class MovieFeeCalculationServiceTest {
     public void timeBaseDiscount() {
         Movie spiderMan = new Movie("Spider-Man: No Way Home", Duration.ofMinutes(90),10, 1);
         Showing showing = new Showing(spiderMan, 5, LocalDateTime.of(LocalDate.now(), LocalTime.of(11, 0)));
-        assertEquals(2.5, movieFeeCalculationService.getDiscountByShowTime(showing));
+        assertEquals(2.50, movieFeeCalculationService.getDiscountByShowTime(showing).doubleValue());
 
         spiderMan = new Movie("Spider-Man: No Way Home", Duration.ofMinutes(90),10, 1);
         showing = new Showing(spiderMan, 5, LocalDateTime.of(LocalDate.now(), LocalTime.of(8, 0)));
-        assertEquals(0, movieFeeCalculationService.getDiscountByShowTime(showing));
+        assertEquals(0, movieFeeCalculationService.getDiscountByShowTime(showing).doubleValue());
 
     }
 
@@ -44,15 +44,15 @@ public class MovieFeeCalculationServiceTest {
 
         Movie spiderMan = new Movie("Spider-Man: No Way Home", Duration.ofMinutes(90),10, 0);
         Showing showing = new Showing(spiderMan, 1, LocalDateTime.of(LocalDate.now(), LocalTime.of(17, 0)));
-        assertEquals(7.0, movieFeeCalculationService.calculateTicketPrice(showing));
+        assertEquals(7.00, movieFeeCalculationService.calculateTicketPrice(showing).doubleValue());
 
         spiderMan = new Movie("Spider-Man: No Way Home", Duration.ofMinutes(90),10, 0);
         showing = new Showing(spiderMan, 2, LocalDateTime.of(LocalDate.now(), LocalTime.of(17, 0)));
-        assertEquals(8.0, movieFeeCalculationService.calculateTicketPrice(showing));
+        assertEquals(8.00, movieFeeCalculationService.calculateTicketPrice(showing).doubleValue());
 
         spiderMan = new Movie("Spider-Man: No Way Home", Duration.ofMinutes(90),10, 0);
         showing = new Showing(spiderMan, 7, LocalDateTime.of(LocalDate.now(), LocalTime.of(17, 0)));
-        assertEquals(9.0, movieFeeCalculationService.calculateTicketPrice(showing));
+        assertEquals(9.00, movieFeeCalculationService.calculateTicketPrice(showing).doubleValue());
 
     }
